dom/contacts/tests/test_permission_denied.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=674720
-->
<head>
  <title>Test for Bug 1081873</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1081873">Mozilla Bug 1081873</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="text/javascript;version=1.8" src="http://mochi.test:8888/tests/dom/contacts/tests/shared.js"></script>
<script class="testbody" type="text/javascript">
"use strict";

SpecialPowers.removePermission("contacts-write", document);
SpecialPowers.removePermission("contacts-read", document);
SpecialPowers.removePermission("contacts-create", document);

function onUnexpectedSuccess() {
  ok(false, "Unexpected success");
  next();
}

function onExpectedError(event) {
  is(event.target.error.name, PERMISSION_DENIED, "Expected PERMISSION_DENIED");
  next();
}

const PERMISSION_DENIED = "PERMISSION_DENIED";

var steps = [
  function() {
    ok(true, "Add contact without permission");
    var req = navigator.mozContacts.save(new mozContact({}));
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = onExpectedError;
  },
  function() {
    ok(true, "Find contact without permission");
    var req = navigator.mozContacts.find({});
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = onExpectedError;
  },
  function() {
    ok(true, "Get all contacts without permission");
    var req = navigator.mozContacts.getAll();
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = onExpectedError;
  },
  function() {
    ok(true, "Remove contact without permission");
    var req = navigator.mozContacts.remove("aId");
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = onExpectedError;
  },
  function() {
    ok(true, "Clear contacts without permission");
    var req = navigator.mozContacts.clear();
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = onExpectedError;
  },
  function() {
    ok(true, "Get revision without permission");
    var req = navigator.mozContacts.getRevision();
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = onExpectedError;
  },
  function() {
    ok(true, "Get count without permission");
    var req = navigator.mozContacts.getCount();
    req.onsuccess = onUnexpectedSuccess;
    req.onerror = function() {
      is(req.error.name, PERMISSION_DENIED, "Expected PERMISSION_DENIED");
      SimpleTest.finish();
    };
  }
];

start_tests();
</script>
</pre>
</body>
</html>
